@@ -6,12 +6,12 @@ showTextFieldDialog({
   String okTitle = "OK",
   String cancelTitle = "Cancel",
   TextInputType type = TextInputType.text,
-  @required action(String text),
+  required action(String text),
 }) {
   TextEditingController _controller = TextEditingController();
   dialog(
       AnimatedContainer(
-          padding: MediaQuery.of(context).viewInsets,
+          padding: MediaQuery.of(context!).viewInsets,
           duration: Duration(milliseconds: 300),
           child: AlertDialog(
               title: Text(title),
