@@ -1,21 +1,21 @@
 part of '../utilities.dart';
 
-Future request(String url, EHttpMethod httpMethod, action(http.Response response), error(http.Response response), {body}) async {
-  final String token = await getString(Constant.token);
+Future request(String url, EHttpMethod httpMethod, action(http.Response? response), error(http.Response? response), {body}) async {
+  final String? token = await getString(Constant.token);
 
-  Map<String, String> headers = {"Content-Type": "application/json", "Authorization": token};
+  Map<String, String?> headers = {"Content-Type": "application/json", "Authorization": token};
 
-  http.Response response;
-  if (httpMethod == EHttpMethod.get) response = await http.get(Uri.parse(url), headers: headers);
-  if (httpMethod == EHttpMethod.post) response = body == null ? await http.post(Uri.parse(url), headers: headers) : await http.post(Uri.parse(url), body: body.toJson(), headers: headers);
-  if (httpMethod == EHttpMethod.put) response = await http.put(Uri.parse(url), body: body.toJson(), headers: headers);
-  if (httpMethod == EHttpMethod.patch) response = await http.patch(Uri.parse(url), body: body.toJson(), headers: headers);
-  if (httpMethod == EHttpMethod.delete) response = await http.delete(Uri.parse(url), headers: headers);
+  http.Response? response;
+  if (httpMethod == EHttpMethod.get) response = await http.get(Uri.parse(url), headers: headers as Map<String, String>?);
+  if (httpMethod == EHttpMethod.post) response = body == null ? await http.post(Uri.parse(url), headers: headers as Map<String, String>?) : await http.post(Uri.parse(url), body: body.toJson(), headers: headers as Map<String, String>?);
+  if (httpMethod == EHttpMethod.put) response = await http.put(Uri.parse(url), body: body.toJson(), headers: headers as Map<String, String>?);
+  if (httpMethod == EHttpMethod.patch) response = await http.patch(Uri.parse(url), body: body.toJson(), headers: headers as Map<String, String>?);
+  if (httpMethod == EHttpMethod.delete) response = await http.delete(Uri.parse(url), headers: headers as Map<String, String>?);
 
   if (body != null)
-    response.completeLog(params: body.toJson());
+    response!.completeLog(params: body.toJson());
   else
-    response.log();
+    response!.log();
 
   if (response.isSuccessful())
     action(response);
@@ -24,40 +24,40 @@ Future request(String url, EHttpMethod httpMethod, action(http.Response response
 }
 
 Future get({
-  @required String url,
-  @required action(http.Response response),
-  @required error(http.Response response),
+  required String url,
+  required action(http.Response? response),
+  required error(http.Response? response),
 }) async =>
     await request(url, EHttpMethod.get, action, error);
 
 Future post({
-  @required String url,
-  @required action(http.Response response),
-  @required error(http.Response response),
+  required String url,
+  required action(http.Response? response),
+  required error(http.Response? response),
   body,
 }) async =>
     await request(url, EHttpMethod.post, action, error, body: body);
 
 Future put({
-  @required String url,
-  @required action(http.Response response),
-  @required error(http.Response response),
+  required String url,
+  required action(http.Response? response),
+  required error(http.Response? response),
   body,
 }) async =>
     await request(url, EHttpMethod.put, action, error, body: body);
 
 Future patch({
-  @required String url,
-  @required action(http.Response response),
-  @required error(http.Response response),
+  required String url,
+  required action(http.Response? response),
+  required error(http.Response? response),
   body,
 }) async =>
     await request(url, EHttpMethod.patch, action, error, body: body);
 
 Future delete({
-  @required String url,
-  @required action(http.Response response),
-  @required error(http.Response response),
+  required String url,
+  required action(http.Response? response),
+  required error(http.Response? response),
 }) async =>
     await request(url, EHttpMethod.delete, action, error);
 
@@ -68,9 +68,9 @@ extension HTTP on http.Response {
 
   bool isServerError() => this.statusCode >= 500 && this.statusCode <= 599 ? true : false;
 
-  log() => print("${this.request.method} - ${this.request.url} - ${this.statusCode} - RESPONSE: ${this.body}");
+  log() => print("${this.request!.method} - ${this.request!.url} - ${this.statusCode} - RESPONSE: ${this.body}");
 
-  completeLog({String params}) => print(
-        "METHOD: ${this.request.method} - URL: ${this.request.url} - STATUS CODE: ${this.statusCode} - RESPONSE: ${this.body} - PARAMS:$params",
+  completeLog({String? params}) => print(
+        "METHOD: ${this.request!.method} - URL: ${this.request!.url} - STATUS CODE: ${this.statusCode} - RESPONSE: ${this.body} - PARAMS:$params",
       );
 }
