@@ -7,13 +7,13 @@ import 'utils/consts.dart';
 import 'widget/render_table.dart';
 
 class DatePicker extends StatefulWidget {
-  final bool isRangeDate;
+  final bool? isRangeDate;
   final bool isSecondDate;
   final startSelectedDate;
   final endSelectedDate;
-  final Function(dynamic) onSelectDate;
-  final Function(String) onConfirmedDate;
-  final Function(String) onChangePicker;
+  final Function(dynamic)? onSelectDate;
+  final Function(String)? onConfirmedDate;
+  final Function(String)? onChangePicker;
 
   DatePicker({this.isRangeDate, this.startSelectedDate = false, this.isSecondDate = false, this.endSelectedDate, this.onChangePicker, this.onSelectDate, this.onConfirmedDate});
 
@@ -22,13 +22,13 @@ class DatePicker extends StatefulWidget {
 }
 
 class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
-  AnimationController controller;
-  Animation<double> animation;
+  late AnimationController controller;
+  late Animation<double> animation;
 
-  Jalali initDate;
-  Jalali startSelectedDate;
-  Jalali endSelectedDate;
-  bool isRangeDate;
+  Jalali? initDate;
+  Jalali? startSelectedDate;
+  Jalali? endSelectedDate;
+  bool? isRangeDate;
 
   bool isSlideForward = true;
 
@@ -90,20 +90,20 @@ class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
   _changeMonth(type) {
     setState(() {
       controller.forward(from: 0);
-      int year = int.parse(initDate.formatter.y);
-      int month = int.parse(initDate.formatter.m);
+      int year = int.parse(initDate!.formatter.y);
+      int month = int.parse(initDate!.formatter.m);
       //int day = int.parse(initDate.formatter.d);
       var newDate = initDate;
       switch (type) {
         case 'prev':
           isSlideForward = true;
 
-          newDate = initDate.copy(month: month > 1 ? month - 1 : 12, year: month == 1 ? year - 1 : year);
+          newDate = initDate!.copy(month: month > 1 ? month - 1 : 12, year: month == 1 ? year - 1 : year);
           break;
         case 'next':
           isSlideForward = false;
 
-          newDate = initDate.copy(month: month < 12 ? month + 1 : 1, year: month == 12 ? year + 1 : year);
+          newDate = initDate!.copy(month: month < 12 ? month + 1 : 1, year: month == 12 ? year + 1 : year);
           break;
         case 'now':
           newDate = Jalali.now();
@@ -175,17 +175,17 @@ class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
               decoration: BoxDecoration(
                 color: Global.color,
               ),
-              child: isRangeDate
+              child: isRangeDate!
                   ? Column(
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: <Widget>[
                         Text(
-                          'از ${fullFormat(startSelectedDate)}',
+                          'از ${fullFormat(startSelectedDate!)}',
                           textAlign: TextAlign.right,
                           style: TextStyle(color: Colors.white),
                         ),
                         Text(
-                          'تا ${fullFormat(endSelectedDate)}',
+                          'تا ${fullFormat(endSelectedDate!)}',
                           textAlign: TextAlign.right,
                           style: TextStyle(color: Colors.white),
                         ),
@@ -195,12 +195,12 @@ class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
                       crossAxisAlignment: CrossAxisAlignment.start,
                       children: <Widget>[
                         Text(
-                          '${startSelectedDate.formatter.yyyy}',
+                          '${startSelectedDate!.formatter.yyyy}',
                           textAlign: TextAlign.right,
                           style: TextStyle(color: Colors.white),
                         ),
                         Text(
-                          '${monthDayFormat(startSelectedDate)}',
+                          '${monthDayFormat(startSelectedDate!)}',
                           textAlign: TextAlign.right,
                           style: TextStyle(color: Colors.white, fontSize: 18, fontWeight: FontWeight.bold),
                         ),
@@ -231,9 +231,9 @@ class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
                               opacity: 1 - animation.value,
                               child: FlatButton(
                                 onPressed: () {
-                                  widget.onChangePicker('year');
+                                  widget.onChangePicker!('year');
                                 },
-                                child: Text(yearMonthNFormat(initDate)),
+                                child: Text(yearMonthNFormat(initDate!)),
                               ),
                             )),
                       ),
@@ -249,7 +249,7 @@ class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
                 Container(
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.center,
-                    children: weekDaysWidget,
+                    children: weekDaysWidget as List<Widget>,
                   ),
                 ),
                 Transform(
@@ -261,7 +261,7 @@ class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
                       startSelectedDate: startSelectedDate,
                       endSelectedDate: endSelectedDate,
                       onSelect: (date) {
-                        widget.onSelectDate(date);
+                        widget.onSelectDate!(date);
                       },
                     ),
                   ),
@@ -280,10 +280,10 @@ class _DatePickerState extends State<DatePicker> with TickerProviderStateMixin {
                     style: TextStyle(fontSize: 16, color: Global.color),
                   ),
                   onPressed: () {
-                    if (isRangeDate) {
-                      widget.onConfirmedDate('${outPutFormat(startSelectedDate)} # ${outPutFormat(endSelectedDate)}');
+                    if (isRangeDate!) {
+                      widget.onConfirmedDate!('${outPutFormat(startSelectedDate!)} # ${outPutFormat(endSelectedDate!)}');
                     } else {
-                      widget.onConfirmedDate('${outPutFormat(startSelectedDate)}');
+                      widget.onConfirmedDate!('${outPutFormat(startSelectedDate!)}');
                     }
                   },
                 ),
