@@ -5,7 +5,7 @@ import 'utils/date.dart';
 
 class TimePicker extends StatefulWidget {
   final initTime;
-  final Function(String) onSelectDate;
+  final Function(String)? onSelectDate;
 
   TimePicker({this.initTime, this.onSelectDate});
 
@@ -14,8 +14,8 @@ class TimePicker extends StatefulWidget {
 }
 
 class _TimePickerState extends State<TimePicker> with TickerProviderStateMixin {
-  AnimationController controller;
-  Animation<double> animation;
+  late AnimationController controller;
+  late Animation<double> animation;
   var initHour;
   var initMinute;
   String changeState = 'hour';
@@ -242,7 +242,7 @@ class _TimePickerState extends State<TimePicker> with TickerProviderStateMixin {
                     style: TextStyle(fontSize: 16, color: isDisable ? Global.color.withOpacity(0.5) : Global.color),
                   ),
                   onPressed: () {
-                    if (!isDisable) widget.onSelectDate('$initHour:$initMinute');
+                    if (!isDisable) widget.onSelectDate!('$initHour:$initMinute');
                   },
                 ),
                 FlatButton(
