@@ -1,10 +1,10 @@
 import 'dart:collection';
 
-Iterable<List<T>> partition<T>(Iterable<T> iterable, int size) {
-  return iterable.isEmpty ? [] : new _Partition<T>(iterable, size);
+Iterable<List<T>?> partition<T>(Iterable<T> iterable, int size) {
+  return (iterable.isEmpty ? [] : new _Partition<T>(iterable, size)) as Iterable<List<T>?>;
 }
 
-class _Partition<T> extends IterableBase<List<T>> {
+class _Partition<T> extends IterableBase<List<T>?> {
   final Iterable<T> _iterable;
   final int _size;
 
@@ -12,18 +12,18 @@ class _Partition<T> extends IterableBase<List<T>> {
     if (_size <= 0) throw new ArgumentError(_size);
   }
 
-  Iterator<List<T>> get iterator => new _PartitionIterator<T>(_iterable.iterator, _size);
+  Iterator<List<T>?> get iterator => new _PartitionIterator<T>(_iterable.iterator, _size);
 }
 
-class _PartitionIterator<T> implements Iterator<List<T>> {
+class _PartitionIterator<T> implements Iterator<List<T>?> {
   final Iterator<T> _iterator;
   final int _size;
-  List<T> _current;
+  List<T>? _current;
 
   _PartitionIterator(this._iterator, this._size);
 
   @override
-  List<T> get current => _current;
+  List<T>? get current => _current;
 
   @override
   bool moveNext() {
