@@ -1,6 +1,6 @@
 part of '../utilities.dart';
 
-launchURL(String url, {bool universal}) async => await canLaunch(url) ? await launch(url, universalLinksOnly: universal) : throw 'Could not launch $url';
+launchURL(String url, {bool? universal}) async => await canLaunch(url) ? await launch(url, universalLinksOnly: universal!) : throw 'Could not launch $url';
 
 launchWhatsApp(String number) async => await launchURL("https://api.whatsapp.com/send?phone=$number");
 
