@@ -1,3 +1,3 @@
 part of '../utilities.dart';
 
-void shareText(String text, {String subject}) => Share.share(text, subject: subject);
+void shareText(String text, {String? subject}) => Share.share(text, subject: subject);
